refactor
A stream cache does not appear to be needed, as YT seems to be caching the streams when swiping thru Shorts.  Streams have authentication time limits as well so it's easier and more reliable to just fetch again if the same video is closed then reopened.

Can use a regular Future as none of the features of CompletableFuture is being used here.  Using an executor future also avoids any potential StrictMode policy violations if the calling YT thread is set to never make network calls.